+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 ----------------
 
 - Use SelectWidget with pat-select2 for Plone 5.
+- Hide all location fields in search form tile when one option in linked search form is pre-selected.
 
 
 1.0 (2017-09-06)